modules/cameraCapture/requirements.txt
azure-iot-device~=2.0.0
requests
paho-mqtt